fix a signed/unsigned compare

+++ miniupnpd/miniupnpd.c
@@ -1,4 +1,4 @@
-/* $Id: miniupnpd.c,v 1.156 2012/05/08 20:41:45 nanard Exp $ */
+/* $Id: miniupnpd.c,v 1.157 2012/05/09 10:41:16 nanard Exp $ */
 /* MiniUPnP project
  * http://miniupnp.free.fr/ or http://miniupnp.tuxfamily.org/
  * (c) 2006-2012 Thomas Bernard
@@ -1430,7 +1430,7 @@ main(int argc, char * * argv)
 		next_pinhole_ts = 0;
 		upnp_clean_expired_pinholes(&next_pinhole_ts);
 		if(next_pinhole_ts &&
-		   timeout.tv_sec >= (next_pinhole_ts - timeofday.tv_sec)) {
+		   timeout.tv_sec >= (int)(next_pinhole_ts - timeofday.tv_sec)) {
 			timeout.tv_sec = next_pinhole_ts - timeofday.tv_sec;
 			timeout.tv_usec = 0;
 		}
